flux error again
