+++ src/components/Details.js
@@ -1,0 +1,11 @@
+import React, { useContext } from "react";
+
+
+const Details = (props) => {
+    //const { store, actions } = useContext(Context);
+    return <div>
+    </div>
+}
+
+export default Details;
+
