+++ src/store/flux.js
@@ -1,6 +1,14 @@
 const getState = ({ getStore, setStore, getActions }) => {
     return {
         store: {
+            user: [],
+            users: [],
+            favorite: [],
+            favorites: [],
+            match: [],
+            product: [],
+            products: [],
+            favorite_list: [],
             products: null,
             product: {
                     title: "",
@@ -10,7 +18,7 @@ const getState = ({ getStore, setStore, getActions }) => {
             },
         actions: {
             getProducts: () => {
-                    fetch("http://localhost:3000/products")
+                    fetch("http://localhost:5000/products")
                     .then(res => res.json())
                     .then(data => setStore({products: data}))
                     .catch(error => console.log(error))
@@ -22,7 +30,7 @@ const getState = ({ getStore, setStore, getActions }) => {
             saveProduct: (e) => {
                 e.preventDefault()
                 const {product} = getStore()
-                fetch("http://localhost:3000/product", {
+                fetch("http://localhost:5000/product", {
                     method: "POST",
                     headers: {
                         "Content-Type":"application/json"
@@ -36,6 +44,47 @@ const getState = ({ getStore, setStore, getActions }) => {
                     autor:""
                 }})
             }
+            addfav: favorite => {
+                const store = getStore();
+                if (store.favorite_list.includes(favorite)){
+                    return console.log("Ya existe este libro como tu favorito")
+                }
+                setStore({favorite_list: [...store.favorite_list, favorite] })
+                return console.log(store.favorite_list)
+            },
+            deleteFav: (index) => {
+                const store = getStore()
+                store.favorite_list.splice(index,1)
+                setStore({favorite_list:store.favorite_list})
+            },
+
+            /*PARA CONSULTA DE LOGGIN Y REGISTRO DE SIGN IN*/
+
+            /*para cuando se inicia sesión LOG IN*/
+            getUser: id => {
+                fetch("" + id)
+                    .then(response => response.json())
+                    .then((result) => {
+                            setStore({user: FALTA LA RUTA DE LA BASE DE DATOS });
+
+                    })
+                    .catch(error => console.log("Error", error));
+            },
+            /*para cuando se hace la consulta de registro si el usurio existe o no (registrado) SIGN IN*/
+            
+            getUsers: () => {
+                fetch("http://localhost:5000/")
+                    .then(response => response.json())
+                    .then(data => setStore({users: RUTA A LA BASE DE DATOS}));
+             },
+            /*CONSULTA A LA BASE DE DATOS DESDE "BUSQUEDA"*/
+             getBooks: ()  => {
+                 fetch("http://localhost:5000/")
+                 .then (response => response.json())
+                 .then ((data)) =>
+                        setStore({ planet: ruta en base de datos}));
+
+             }
         }
     };
 };
